+++ qiskit/transpiler/passes/swap_mapper.py
@@ -79,8 +79,8 @@ class SwapMapper(TransformationPass):
                     farthest_physical = path[-1]['name'][1]
                     new_dag.add_gate_data(self.swap_basis_element, self.swap_data)
                     new_dag.apply_operation_back(self.swap_basis_element,
-                                                 [self.layout[closest_physical],
-                                                  self.layout[farthest_physical]])
+                                                 [('q', closest_physical),
+                                                  ('q', farthest_physical)])
                     self.layout.swap(farthest_physical, closest_physical)
                 new_dag.add_dag_at_the_end(subdag, self.layout)
         return new_dag