+++ test/python/transpiler/test_swap_mapper.py
@@ -67,7 +67,7 @@ class TestSwapMapper(QiskitTestCase):
         after_dag = pass_.run(dag)
 
         self.assertEqual(before, after_dag.qasm())
-        self.assertDictEqual(pass_.layout.get_logical(), {('q', 0): 0,
+        self.assertDictEqual(pass_.layout.get_bits(), {('q', 0): 0,
                                                           ('q', 1): 1,
                                                           ('q', 2): 2})
 
@@ -92,7 +92,7 @@ class TestSwapMapper(QiskitTestCase):
         after_dag = pass_.run(dag)
 
         self.assertEqual(before, after_dag.qasm())
-        self.assertDictEqual(pass_.layout.get_logical(), {('q', 0): 0,
+        self.assertDictEqual(pass_.layout.get_bits(), {('q', 0): 0,
                                                           ('q', 1): 1,
                                                           ('q', 2): 2,
                                                           ('q', 3): 3})
@@ -127,34 +127,66 @@ class TestSwapMapper(QiskitTestCase):
         after_dag = pass_.run(dag)
 
         self.assertEqual(expected, after_dag.qasm())
-        self.assertDictEqual(pass_.layout.get_logical(), {('q', 0): 2, ('q', 1): 1, ('q', 2): 0})
+        self.assertDictEqual(pass_.layout.get_bits(), {('q', 0): 2, ('q', 1): 1, ('q', 2): 0})
 
-    def test_far_swap(self):
-        """ A far swap that affects coming CXs.
-         qr0:--(+)---.--
-                |    |
-         qr1:---|----|--
-                |    |
-         qr2:---|----|--
-                |    |
-         qr3:---.---(+)-
+    def test_keep_layout(self):
+        """After a swap, the following gates also change the wires.
+         qr0:---.--------
+                |
+         qr1:---|--------
+                |
+         qr2:--(+)--[U]--
 
-         Coupling map: [0]--[1]--[2]--[3]
+         Coupling map: [0]--[1]--[2]
+
+         qr0:------.---------
+                   |
+         qr1:--X--(+)--[U1]--
+               |
+         qr2:--X-------------
         """
-        coupling = Coupling({0: [1], 1: [2], 2: [3]})
-        dag = TestSwapMapper.create_dag([('CX', [('qr', 0), ('qr', 3)]),
-                                         ('CX', [('qr', 3), ('qr', 0)])])
+        coupling = Coupling({1: [0, 2]})
+        dag = TestSwapMapper.create_dag([('CX', [('qr', 0), ('qr', 2)]),
+                                         ('U', [('qr', 2)])])
+
         expected = '\n'.join(["OPENQASM 2.0;",
                               "qreg q[3];",
                               "opaque swap a,b;",
-                              "CX q[0],q[1];",
-                              "swap q[0],q[2];",
+                              "swap q[1],q[2];",
                               "CX q[1],q[0];"]) + '\n'
         expected = '\n'.join([])
         pass_ = SwapMapper(coupling)
         after_dag = pass_.run(dag)
 
-        # self.assertEqual(expected, after_dag.qasm())
+        self.assertEqual(expected, after_dag.qasm())
+        # self.assertDictEqual(pass_.layout.get_bits(), {('q', 0): 2, ('q', 1): 1, ('q', 2): 0})
+
+    # def test_far_swap(self):
+    #     """ A far swap that affects coming CXs.
+    #      qr0:--(+)---.--
+    #             |    |
+    #      qr1:---|----|--
+    #             |    |
+    #      qr2:---|----|--
+    #             |    |
+    #      qr3:---.---(+)-
+    #
+    #      Coupling map: [0]--[1]--[2]--[3]
+    #     """
+    #     coupling = Coupling({0: [1], 1: [2], 2: [3]})
+    #     dag = TestSwapMapper.create_dag([('CX', [('qr', 0), ('qr', 3)]),
+    #                                      ('CX', [('qr', 3), ('qr', 0)])])
+    #     expected = '\n'.join(["OPENQASM 2.0;",
+    #                           "qreg q[3];",
+    #                           "opaque swap a,b;",
+    #                           "CX q[0],q[1];",
+    #                           "swap q[0],q[2];",
+    #                           "CX q[1],q[0];"]) + '\n'
+    #     expected = '\n'.join([])
+    #     pass_ = SwapMapper(coupling)
+    #     after_dag = pass_.run(dag)
+    #
+    #     # self.assertEqual(expected, after_dag.qasm())
 
 
 if __name__ == '__main__':