+++ CHANGELOG.rst
@@ -43,6 +43,7 @@ Added
   QuantumCircuit class. (#1172)
 - New methods in QuantumCircuit for common circuit metrics:
   `size()`, `depth()`, `width()`, `count_ops()`, `num_tensor_factors()` (#1285)
+- Added a `Layout` object (#1313)
 
 Changed
 """""""